Fix tests to work with 3.x
Version 3.x of reselect takes the input argument equality into account
when deciding whether to recompute. Unfortunately, some of the tests
mutated the value passed to the selector, so the selector wouldn't
update when it should have. This commit changes the updates to be
"immutable" and the test cases work once again.